+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/Settings.kt
@@ -246,6 +246,11 @@ abstract class Settings {
      * Setting to control whether Global Privacy Control isenabled.
      */
     open var globalPrivacyControlEnabled: Boolean by UnsupportedSetting()
+
+    /**
+     * Setting to control the email tracker blocking feature in the private browsing mode.
+     */
+    open var emailTrackerBlockingPrivateBrowsing: Boolean by UnsupportedSetting()
 }
 
 /**
@@ -295,6 +300,7 @@ data class DefaultSettings(
     override var queryParameterStrippingPrivateBrowsing: Boolean = false,
     override var queryParameterStrippingAllowList: String = "",
     override var queryParameterStrippingStripList: String = "",
+    override var emailTrackerBlockingPrivateBrowsing: Boolean = false,
 ) : Settings()
 
 class UnsupportedSetting<T> {